requirement.txt
streamlit>=1.32
openai>=1.40.0
faiss-cpu
numpy
requests
python-dotenv
